test: extract telemetry and logging tests into their own files
