+++ packages/cli-repl/src/setup-analytics.spec.ts
@@ -1,0 +1,74 @@
+import { expect } from 'chai';
+import os from 'os';
+import {
+  NoopAnalytics,
+  ThrottledAnalytics,
+  ToggleableAnalytics,
+} from '@mongosh/logging';
+import { setupTelemetryAnalytics } from './setup-analytics';
+
+describe('setupTelemetryAnalytics', function () {
+  const metadataPath = os.tmpdir();
+  // A fetch stub; these tests never actually track()/send, they only inspect
+  // how the analytics sink is constructed.
+  const fetch = () => Promise.resolve(new Response());
+
+  let savedEnvEndpoint: string | undefined;
+  beforeEach(function () {
+    savedEnvEndpoint = process.env.MONGOSH_TELEMETRY_ENDPOINT;
+    delete process.env.MONGOSH_TELEMETRY_ENDPOINT;
+  });
+  afterEach(function () {
+    if (savedEnvEndpoint === undefined) {
+      delete process.env.MONGOSH_TELEMETRY_ENDPOINT;
+    } else {
+      process.env.MONGOSH_TELEMETRY_ENDPOINT = savedEnvEndpoint;
+    }
+  });
+
+  function setup(
+    params: Partial<Parameters<typeof setupTelemetryAnalytics>[0]> = {}
+  ) {
+    return setupTelemetryAnalytics({
+      configuredTelemetryEndpoint: '',
+      fetch: fetch as any,
+      metadataPath,
+      ...params,
+    });
+  }
+
+  it('returns a no-op sink when no endpoint is configured', function () {
+    const { analytics, telemetryEndpoint } = setup();
+    expect(telemetryEndpoint).to.equal('');
+    expect(analytics).to.be.instanceOf(ToggleableAnalytics);
+    // No endpoint -> nothing to send to. Telemetry is not disabled here;
+    // events are still logged locally, they just have no destination.
+    expect(analytics._target).to.be.instanceOf(NoopAnalytics);
+  });
+
+  it('creates a telemetry client when an endpoint is configured via user config', function () {
+    const { analytics, telemetryEndpoint } = setup({
+      configuredTelemetryEndpoint: 'https://config.example/events',
+    });
+    expect(telemetryEndpoint).to.equal('https://config.example/events');
+    expect(analytics._target).to.be.instanceOf(ThrottledAnalytics);
+  });
+
+  it('uses MONGOSH_TELEMETRY_ENDPOINT over the configured default', function () {
+    process.env.MONGOSH_TELEMETRY_ENDPOINT = 'https://env.example/events';
+    const { telemetryEndpoint, analytics } = setup({
+      configuredTelemetryEndpoint: 'https://config.example/events',
+    });
+    expect(telemetryEndpoint).to.equal('https://env.example/events');
+    expect(analytics._target).to.be.instanceOf(ThrottledAnalytics);
+  });
+
+  it('is disabled when every source resolves to an empty endpoint', function () {
+    process.env.MONGOSH_TELEMETRY_ENDPOINT = '';
+    const { telemetryEndpoint, analytics } = setup({
+      configuredTelemetryEndpoint: '',
+    });
+    expect(telemetryEndpoint).to.equal('');
+    expect(analytics._target).to.be.instanceOf(NoopAnalytics);
+  });
+});
