+++ packages/logging/src/logging-and-telemetry.spec.ts
@@ -181,6 +181,7 @@ describe('MongoshLoggingAndTelemetry', function () {
             mongosh_version: '1.0.0',
             ai_agent: undefined,
             session_id: logId,
+            device_id: deviceId,
             is_localhost: true,
             is_atlas: false,
             atlas_hostname: null,
@@ -232,6 +233,7 @@ describe('MongoshLoggingAndTelemetry', function () {
             mongosh_version: '1.0.0',
             ai_agent: undefined,
             session_id: logId,
+            device_id: deviceId,
             is_localhost: false,
             is_atlas: true,
             atlas_hostname: 'test-data-sets-00-02-a011bb.mongodb.net',
@@ -288,6 +290,7 @@ describe('MongoshLoggingAndTelemetry', function () {
         mongosh_version: '1.0.0',
         ai_agent: undefined,
         session_id: logId,
+        device_id: deviceId,
         is_atlas: true,
         is_atlas_url: true,
         is_local_atlas: false,
@@ -875,6 +878,7 @@ describe('MongoshLoggingAndTelemetry', function () {
             mongosh_version: '1.0.0',
             ai_agent: undefined,
             session_id: logId,
+            device_id: deviceId,
             is_interactive: true,
             commands_repl: undefined,
             commands_rc: { 'ShellApi.use': 1 },
@@ -1017,6 +1021,7 @@ describe('MongoshLoggingAndTelemetry', function () {
             mongosh_version: '1.0.0',
             ai_agent: undefined,
             session_id: logId,
+            device_id: deviceId,
             is_interactive: true,
             commands_repl: undefined,
             commands_rc: {
@@ -1088,6 +1093,7 @@ describe('MongoshLoggingAndTelemetry', function () {
             mongosh_version: '1.0.0',
             ai_agent: undefined,
             session_id: logId,
+            device_id: deviceId,
             is_interactive: true,
             commands_repl: undefined,
             commands_rc: undefined,
@@ -1298,6 +1304,7 @@ describe('MongoshLoggingAndTelemetry', function () {
             mongosh_version: '1.0.0',
             ai_agent: undefined,
             session_id: logId,
+            device_id: deviceId,
             is_localhost: true,
             is_atlas: false,
             atlas_hostname: null,
