+++ packages/logging/src/telemetry-events.ts
@@ -57,6 +57,8 @@ export interface NewConnectionEvent {
     ai_agent: string | undefined;
     /** Unique identifier for the current mongosh session. */
     session_id: string;
+    /** A stable machine identifier; `"unknown"` if unavailable. */
+    device_id: string;
     /** Whether the server is an Atlas deployment. */
     is_atlas: boolean;
     /** Whether the connection URI is an Atlas URL. */
@@ -130,6 +132,8 @@ export interface SessionEndedEvent {
     ai_agent: string | undefined;
     /** Unique identifier for the current mongosh session. */
     session_id: string;
+    /** A stable machine identifier; `"unknown"` if unavailable. */
+    device_id: string;
     /** Whether mongosh was started in interactive (REPL) mode. */
     is_interactive: boolean;
     /** Map of command keys with counts from REPL evaluations. */
